+++ ProgramAssignment1/STEPINC_MIN2MAX.cpp
@@ -13,7 +13,7 @@
 double STEPINC_MIN2MAX::AlgorithmUse(double min_data, double max_data, double step_data, double current_data)
 {
     // Rand
-    const int generated = GenerateNewRandom(0, step_data);
+    const double generated = GenerateNewRandom(0, step_data);
     
     if( current_data + generated < max_data )
         return current_data += generated;
